answer-service/pyproject.toml
[tool.poetry]
name = "answer-service"
version = "0.1.0"
description = "Conversation and answer generation testing service using Vertex AI conversation capabilities"
authors = ["Genesis <genesis@local>"]
readme = "README.md"
packages = [{include = "answer_service", from = "src"}]

[tool.poetry.scripts]
answer-service = "answer_service.main:main"

[tool.poetry.dependencies]
python = "^3.13"
# Genesis shared utilities will be available once published to PyPI
# shared-core = "^0.1.0"  # ConfigLoader, get_logger, retry, health checks, etc.
click = "^8.1.7"
rich = "^13.7.0"
google-cloud-discoveryengine = "^0.11.0"  # Vertex AI conversation capabilities

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"
ruff = "^0.1.6"
# Note: Genesis CLI is automatically installed during bootstrap

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["answer_service"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
pythonpath = ["src"]
addopts = [
    "--cov=answer_service",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80"
]

[tool.coverage.run]
omit = [
    "*/service_real.py",  # Exclude real Google Cloud API implementation from coverage
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
